Removed the separate codeplay_oneapi.py file and merged class into amd package.py.
